// packages/core/src/generated/types.ts
export interface BatchInvoiceType {
  /**
   * Sequence number of the modification document within the batch
   *
   * A módosító okirat sorszáma a kötegen belül
   */
  batchIndex: number;
  /**
   * Data of a single invoice or modification document
   *
   * Egy számla vagy módosító okirat adatai
   */
  invoice: InvoiceType;
}
// ...
export interface CustomerInfoType {
  /**
   * Customers status by VAT
   *
   * Vevő ÁFA szerinti státusza
   */
  customerVatStatus: CustomerVatStatusType;
  /**
   * VAT subjectivity data of the customer
   *
   * A vevő ÁFA alanyisági adatai
   */
  customerVatData?: CustomerVatDataType;
  /**
   * Name of the customer
   *
   * A vevő neve
   */
  customerName?: string;
  /**
   * Address of the customer
   *
   * A vevő címe
   */
  customerAddress?: AddressType;
  /**
   * Bank account number of the customer
   *
   * Vevő bankszámlaszáma
   */
  customerBankAccountNumber?: string;
}
// ...
export interface DetailedReasonType {
  /**
   * Case notation with code
   *
   * Az eset leírása kóddal
   */
  case: string;
  /**
   * Case notation with text
   *
   * Az eset leírása szöveggel
   */
  reason: string;
}
// ...
export interface InvoiceData {
  /**
   * Sequential number of the original invoice or modification document - section
   * 169 (b) or section 170 (1) b) of the VAT law
   *
   * Számla vagy módosító okirat sorszáma - ÁFA tv. 169. § b) vagy 170. § (1)
   * bek. b) pont
   */
  invoiceNumber: string;
  /**
   * Date of issue of the invoice or the modification document - section 169 (a)
   * of the VAT law, section 170 (1) a) of the VAT law
   *
   * Számla vagy módosító okirat kelte - ÁFA tv. 169. § a), ÁFA tv. 170. § (1)
   * bek. a)
   */
  invoiceIssueDate: string;
  /**
   * Indicates whether the data exchange is identical with the invoice (the
   * invoice does not contain any more data)
   *
   * Jelöli, ha az adatszolgáltatás maga a számla (a számlán nem szerepel több
   * adat)
   */
  completenessIndicator: boolean;
  /**
   * A common type to describe invoice information
   *
   * Számlaadatok leírására szolgáló közös típus
   */
  invoiceMain: InvoiceMainType;
}
// ...
export interface InvoiceHeadType {
  /**
   * Data related to the issuer of the invoice (supplier)
   *
   * Számla kibocsátó (eladó) adatai
   */
  supplierInfo: SupplierInfoType;
  /**
   * Data related to the customer
   *
   * Vevő adatai
   */
  customerInfo?: CustomerInfoType;
  /**
   * Data related to the fiscal representative
   *
   * Pénzügyi képviselő adatai
   */
  fiscalRepresentativeInfo?: FiscalRepresentativeType;
  /**
   * Invoice detail adata
   *
   * Számla részletező adatok
   */
  invoiceDetail: InvoiceDetailType;
}
// ...
export interface InvoiceMainType {
  /**
   * Data of a single invoice or modification document
   *
   * Egy számla vagy módosító okirat adatai
   */
  invoice?: InvoiceType;
  /**
   * Data of a batch of modification documents
   *
   * Kötegelt módosító okirat adatai
   */
  batchInvoice?: BatchInvoiceType[];
}
// ...
export interface InvoiceReferenceType {
  /**
   * Sequence number of the original invoice, on which the modification occurs -
   * section 170 (1) c) of the VAT law
   *
   * Az eredeti számla sorszáma, melyre a módosítás vonatkozik - ÁFA tv. 170. §
   * (1) c)
   */
  originalInvoiceNumber: string;
  /**
   * Indicates whether the modification references to an original invoice which
   * is not and will not be exchanged
   *
   * Annak jelzése, hogy a módosítás olyan alapszámlára hivatkozik, amelyről nem
   * történt és nem is fog történni adatszolgáltatás
   */
  modifyWithoutMaster: boolean;
  /**
   * The unique sequence number referring to the original invoice
   *
   * A számlára vonatkozó módosító okirat egyedi sorszáma
   */
  modificationIndex: number;
}

/**
 * Processing status of the invoice
 *
 * A számla feldolgozási státusza
 */
export type InvoiceStatusType = 'RECEIVED' | 'PROCESSING' | 'SAVED' | 'DONE' | 'ABORTED';

/**
 * Data of a single invoice or modification document
 *
 * Egy számla vagy módosító okirat adatai
 */
export interface InvoiceType {
  /**
   * Modification or cancellation data
   *
   * A módosítás vagy érvénytelenítés adatai
   */
  invoiceReference?: InvoiceReferenceType;
  /**
   * Data concerning the whole invoice
   *
   * A számla egészét jellemző adatok
   */
  invoiceHead: InvoiceHeadType;
  /**
   * Product/service data appearing on the invoice
   *
   * A számlán szereplő tételek adatai
   */
  invoiceLines?: LinesType;
  /**
   * Summary data of product charges
   *
   * Termékdíjjal kapcsolatos összesítő adatok
   *
   * @remarks At most 2 entries.
   */
  productFeeSummary?: ProductFeeSummaryType[];
  /**
   * Summary data according to VAT law
   *
   * Az ÁFA törvény szerinti összesítő adatok
   */
  invoiceSummary: SummaryType;
}
// ...
export interface LinesType {
  /**
   * Indicates whether the data exchange contains merged line data due to size
   * reduction
   *
   * Jelöli, ha az adatszolgáltatás méretcsökkentés miatt összevont soradatokat
   * tartalmaz
   */
  mergedItemIndicator: boolean;
  /**
   * Product / service item
   *
   * Termék/szolgáltatás tétel
   */
  line: LineType[];
}
// ...
export type MarginSchemeType = 'TRAVEL_AGENCY' | 'SECOND_HAND' | 'ARTWORK' | 'ANTIQUES';
// ...
export interface ProductFeeSummaryType {
  /**
   * Indicating whether the the product fee summary concerns refund or deposit
   *
   * Annak jelzése, hogy a termékdíj összesítés visszaigénylésre (REFUND) vagy
   * raktárba történő beszállításra (DEPOSIT) vonatkozik
   */
  productFeeOperation: ProductFeeOperationType;
  /**
   * Product charges data
   *
   * Termékdíj adatok
   */
  productFeeData: ProductFeeDataType[];
  /**
   * Aggregate product charges
   *
   * Termékdíj összesen
   */
  productChargeSum: string;
  /**
   * Data of the document verifying the declaration submitted on the product fee
   * according to the Paragraph (3), Section 13 and the Paragraph (3) Section 25
   * of the Act LXXXV of 2011
   *
   * A termékdíj bevallását igazoló dokumentum adatai a 2011. évi LXXXV. tv. 13.
   * § (3) szerint és a 25. § (3) szerint
   */
  paymentEvidenceDocumentData?: PaymentEvidenceDocumentDataType;
}
// ...
export interface SummaryByVatRateType {
  /**
   * Marking the tax rate or the fact of tax exemption
   *
   * Adómérték vagy adómentesség jelölése
   */
  vatRate: VatRateType;
  /**
   * Net data of given tax rate
   *
   * Adott adómértékhez tartozó nettó adatok
   */
  vatRateNetData: VatRateNetDataType;
  /**
   * VAT data of given tax rate
   *
   * Adott adómértékhez tartozó ÁFA adatok
   */
  vatRateVatData: VatRateVatDataType;
  /**
   * Gross data of given tax rate
   *
   * Adott adómértékhez tartozó bruttó adatok
   */
  vatRateGrossData?: VatRateGrossDataType;
}

/**
 * Gross data of the invoice summary
 *
 * A számla összesítő bruttó adatai
 */
export interface SummaryGrossDataType {
  /**
   * Gross amount of the invoice expressed in the currency of the invoice
   *
   * A számla bruttó összege a számla pénznemében
   */
  invoiceGrossAmount: string;
  /**
   * Gross amount of the invoice expressed in HUF
   *
   * A számla bruttó összege forintban
   */
  invoiceGrossAmountHUF: string;
}

/**
 * Calculation of invoice totals (not simplified invoice)
 *
 * Számla összesítés (nem egyszerűsített számla esetén)
 */
export interface SummaryNormalType {
  /**
   * Calculation of invoice totals per VAT rates
   *
   * Összesítés ÁFA-mérték szerint
   */
  summaryByVatRate: SummaryByVatRateType[];
  /**
   * Net amount of the invoice expressed in the currency of the invoice
   *
   * A számla nettó összege a számla pénznemében
   */
  invoiceNetAmount: string;
  /**
   * Net amount of the invoice expressed in HUF
   *
   * A számla nettó összege forintban
   */
  invoiceNetAmountHUF: string;
  /**
   * VAT amount of the invoice expressed in the currency of the invoice
   *
   * A számla ÁFA összege a számla pénznemében
   */
  invoiceVatAmount: string;
  /**
   * VAT amount of the invoice expressed in HUF
   *
   * A számla ÁFA összege forintban
   */
  invoiceVatAmountHUF: string;
}

/**
 * Calculation of simplified invoice totals
 *
 * Egyszerűsített számla összesítés
 */
export interface SummarySimplifiedType {
  /**
   * Marking the tax rate or the fact of tax exemption
   *
   * Adómérték vagy adómentesség jelölése
   */
  vatRate: VatRateType;
  /**
   * The gross amount of the sale or service for the given tax amount in the
   * currency of the invoice
   *
   * Az adott adótartalomhoz tartozó értékesítés vagy szolgáltatásnyújtás bruttó
   * összege a számla pénznemében
   */
  vatContentGrossAmount: string;
  /**
   * The gross amount of the sale or service for the given tax amount in HUF
   *
   * Az adott adótartalomhoz tartozó értékesítés vagy szolgáltatásnyújtás bruttó
   * összege forintban
   */
  vatContentGrossAmountHUF: string;
}

/**
 * Data of calculation of invoice totals
 *
 * Számla összesítés adatai
 *
 * @remarks Exactly one of `summaryNormal`, `summarySimplified` may be set.
 */
export interface SummaryType {
  /**
   * Calculation of invoice totals (not simplified invoice)
   *
   * Számla összesítés (nem egyszerűsített számla esetén)
   */
  summaryNormal?: SummaryNormalType;
  /**
   * Calculation of simplified invoice totals
   *
   * Egyszerűsített számla összesítés
   */
  summarySimplified?: SummarySimplifiedType[];
  /**
   * Gross data of the invoice summary
   *
   * A számla összesítő bruttó adatai
   */
  summaryGrossData?: SummaryGrossDataType;
}
// ...
export interface VatAmountMismatchType {
  /**
   * VAT rate, VAT content
   *
   * Adómérték, adótartalom
   */
  vatRate: string;
  /**
   * Case notation with code
   *
   * Az eset leírása kóddal
   */
  case: string;
}

/**
 * Gross data of given tax rate
 *
 * Adott adómértékhez tartozó bruttó adatok
 */
export interface VatRateGrossDataType {
  /**
   * Gross amount of sales or service delivery under a given tax rate expressed
   * in the currency of the invoice
   *
   * Az adott adómértékhez tartozó értékesítés vagy szolgáltatásnyújtás bruttó
   * összege a számla pénznemében
   */
  vatRateGrossAmount: string;
  /**
   * Gross amount of sales or service delivery under a given tax rate expressed
   * in HUF
   *
   * Az adott adómértékhez tartozó értékesítés vagy szolgáltatásnyújtás bruttó
   * összege forintban
   */
  vatRateGrossAmountHUF: string;
}

/**
 * Net data of given tax rate
 *
 * Adott adómértékhez tartozó nettó adatok
 */
export interface VatRateNetDataType {
  /**
   * Net amount of sales or service delivery under a given tax rate expressed in
   * the currency of the invoice
   *
   * Az adott adómértékhez tartozó értékesítés vagy szolgáltatásnyújtás nettó
   * összege a számla pénznemében
   */
  vatRateNetAmount: string;
  /**
   * Net amount of sales or service delivery under a given tax rate expressed in
   * HUF
   *
   * Az adott adómértékhez tartozó értékesítés vagy szolgáltatásnyújtás nettó
   * összege forintban
   */
  vatRateNetAmountHUF: string;
}

/**
 * Marking tax rate or tax exempt supply
 *
 * Az adómérték vagy az adómentes értékesítés jelölése
 *
 * @remarks Exactly one of `vatPercentage`, `vatContent`, `vatExemption`, `vatOutOfScope`, `vatDomesticReverseCharge`, `marginSchemeIndicator`, `vatAmountMismatch`, `noVatCharge` may be set.
 */
export interface VatRateType {
  /**
   * Applied tax rate - section 169 (j) of the VAT law
   *
   * Az alkalmazott adó mértéke - ÁFA tv. 169. § j)
   */
  vatPercentage?: string;
  /**
   * VAT content in case of simplified invoice
   *
   * ÁFA tartalom egyszerűsített számla esetén
   */
  vatContent?: string;
  /**
   * Marking tax exemption - section 169 (m) of the VAT law
   *
   * Az adómentesség jelölése - ÁFA tv. 169. § m)
   */
  vatExemption?: DetailedReasonType;
  /**
   * Out of scope of the VAT law
   *
   * Az ÁFA törvény hatályán kívüli
   */
  vatOutOfScope?: DetailedReasonType;
  /**
   * Marking the national is reverse charge taxation - section 142 of the VAT law
   *
   * A belföldi fordított adózás jelölése - ÁFA tv. 142. §
   */
  vatDomesticReverseCharge?: boolean;
  /**
   * Marking the margin-scheme taxation as per section 169 (p)(q)
   *
   * Különbözet szerinti szabályozás jelölése - ÁFA tv. 169. § p) q)
   */
  marginSchemeIndicator?: MarginSchemeType;
  /**
   * Different cases of mismatching tax base and levied tax
   *
   * Adóalap és felszámított adó eltérésének esetei
   */
  vatAmountMismatch?: VatAmountMismatchType;
  /**
   * No VAT charged under Section 17
   *
   * Nincs felszámított áfa a 17. § alapján
   */
  noVatCharge?: boolean;
}

/**
 * VAT data of given tax rate
 *
 * Adott adómértékhez tartozó ÁFA adatok
 */
export interface VatRateVatDataType {
  /**
   * VAT amount of sales or service delivery under a given tax rate expressed in
   * the currency of the invoice
   *
   * Az adott adómértékhez tartozó értékesítés vagy szolgáltatásnyújtás ÁFA
   * összege a számla pénznemében
   */
  vatRateVatAmount: string;
  /**
   * VAT amount of sales or service delivery under a given tax rate expressed in
   * HUF
   *
   * Az adott adómértékhez tartozó értékesítés vagy szolgáltatásnyújtás ÁFA
   * összege forintban
   */
  vatRateVatAmountHUF: string;
}